+++ install-stubs/Orchid/Screens/User/UserEditScreen.php
@@ -4,17 +4,16 @@ declare(strict_types=1);
 
 namespace App\Orchid\Screens\User;
 
-use Orchid\Screen\Link;
-use Orchid\Screen\Screen;
-use Orchid\Screen\Layouts;
-use Illuminate\Http\Request;
-use Orchid\Platform\Models\Role;
-use Orchid\Platform\Models\User;
-use Orchid\Support\Facades\Alert;
-use Orchid\Screen\Fields\Password;
-use Illuminate\Support\Facades\Auth;
 use App\Orchid\Layouts\User\UserEditLayout;
 use App\Orchid\Layouts\User\UserRoleLayout;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Orchid\Platform\Models\User;
+use Orchid\Screen\Fields\Password;
+use Orchid\Screen\Layouts;
+use Orchid\Screen\Link;
+use Orchid\Screen\Screen;
+use Orchid\Support\Facades\Alert;
 
 class UserEditScreen extends Screen
 {
@@ -41,10 +40,11 @@ class UserEditScreen extends Screen
      */
     public function query(User $user): array
     {
+        $user->load(['roles']);
+
         return [
             'user'       => $user,
             'permission' => $user->getStatusPermission(),
-            'roles'      => $user->getStatusRoles(),
         ];
     }
 
@@ -109,7 +109,7 @@ class UserEditScreen extends Screen
     public function save(User $user, Request $request)
     {
         $permissions = $request->get('permissions', []);
-        $roles = Role::whereIn('slug', $request->get('roles', []))->get();
+        $roles = $request->input('user.roles', []);
 
         foreach ($permissions as $key => $value) {
             unset($permissions[$key]);
@@ -121,8 +121,9 @@ class UserEditScreen extends Screen
             ->fill([
                 'permissions' => $permissions,
             ])
-            ->replaceRoles($roles)
             ->save();
+
+        $user->roles()->syncWithoutDetaching($roles);
 
         Alert::info(__('User was saved'));